pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mifud"
version = "0.1.0"
description = "Система предзаказа еды и бронирования столов в столовой университета"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-migrate>=4.0",
    "flask-login>=0.6",
    "flask-wtf>=1.2",
    "python-dotenv>=1.0",
    "wtforms>=3.1",
    "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-flask>=1.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
